Simo changes again

+++ SellukeittoSovellus/SellukeittoSovellus/Controller.cs
@@ -21,11 +21,11 @@ namespace SellukeittoSovellus
         #endregion
 
         #region CLASS VARIABLES
-        
-        public int Cooking_time;
-        public int Cooking_temperature;
-        public int Cooking_pressure;
-        public int Impregnation_time;
+
+        public double Cooking_time;
+        public double Cooking_temperature;
+        public double Cooking_pressure;
+        public double Impregnation_time;
 
         public int State; // Controller state
